// generated/src/app/api-models/subsc-id-change-job-setup.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class SubscIdChangeJobSetup {
  seqSubidId: number;
  jobId: string;
  requestUser: string;
  requestDate: Date;
  action: number;
  status: number;
  requestType: string;
  subsChangeJobComments: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}